+++ potfill/src/main/java/com/potfill/admin/district/model/Complaint.java
@@ -1,4 +1,4 @@
-package com.potfill.admin.managerdistrict.model;
+package com.potfill.admin.district.model;
 
 import java.util.Date;
