+++ potfill/src/main/resources/mappers/admin/district/DistrictMapper.xml
@@ -1,10 +1,10 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
     "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper
-	namespace="com.potfill.admin.managerdistrict.dao.ComplaintRepository">
+	namespace="com.potfill.admin.district.dao.DistrictRepository">
 
 	<resultMap id="ComplaintResultMap"
-		type="com.potfill.admin.managerdistrict.model.Complaint">
+		type="com.potfill.admin.district.model.Complaint">
 		<result property="complaintId" column="COMPLAINT_ID" />
 		<result property="reporterName" column="REPORTER_NAME" />
 		<result property="reporterNumber" column="REPORTER_NUMBER" />
@@ -101,7 +101,7 @@
 
 	<!-- 5. 최근 7일 일자별 건수 -->
 	<resultMap id="dailyCountMap"
-		type="com.potfill.admin.managerdistrict.model.DailyCount">
+		type="com.potfill.admin.district.model.DailyCount">
 		<result property="date" column="CREATED_DATE" />
 		<result property="newCount" column="NEWCOUNT" />
 		<result property="completedCount" column="COMPLETEDCOUNT" />
